+++ langchain4j-alloydb-pg/src/main/java/dev/langchain4j/engine/MetadataColumn.java
@@ -1,4 +1,4 @@
-package dev.langchain4j.store.alloydb.engine;
+package dev.langchain4j.engine;
 
 public class MetadataColumn {
